Motor PWM: OK for a while
At initialisation Motor objects has got a config list / tuple, and four pins.
The pins are registered as motor driver pins (immutable setting).
The tuple contains all info for the proper motor handling:
1. Duty factor (for the left motor) and borders (min. and max. in duty)
2. PWM settings for the left motor: frequency (immutable) and initial duty.
3. Timer setting for the right motor: period (~freq) and sleep (~duty).
   (both immutable)

The initialisation modifies the first part of this tuple:
- Duty factor extracted into a dedicated variable.
- From the given minimum duty and initial duty calculates the minimum duty factor.
- From the given maximum duty and initial duty calculates the maximum duty factor.

The duty factor is settable: setDutyFactor()
This factor affects directly the duty of the left motor.