Improved _startOrStop(), -0.1 as under def., motor callback
_startOrStop() sets callback for motor with _turtleEndSignal, and checks the _toPlay bytearray more safely
(with the help of remaining). If the loop or function is unfinished, it stops the execution.

_functionPosition array indicates with -1 the undefined and with -0.1 the currently edited (not defined yet)
functions. 0, 1 and so on are for indices. (_functionPosition indicates the first curly brace of function.)

Motor class has got a _callback tuple: (callback function, isTemporary flag) and it executes the given
function (if there is any) if it has no more task.