Add wrapper feedback.py
This file is created by the inisetup.py at the first run. It is a general python file
that users can reach and modify. This nature gives the opportunity to adapt to various
I2C chips. There will be other implemented and interpreted 'ubot_' modules for general
ones ('ubot_lsm303.py', ...), and the feedback.py will import and wrap the choosen one.
(Or in extreme cases it will implement the necessary logic.)