WebServer with uselect, feedback pins changed
The modification of webserver violates the 'permanent boot'.
Until this point at startup, ubot_boot.py was the main loop,
because at the end, this called webserver.start() which init
an infinite while loop with socket.accept() inside.

The uselect module with a timer replaces this logic, so after
executing ubot_boot.py, the ESP starts the execution of
ubot_debug.py. Temporary patch: A busy waiting for loop
at the end of the ubot_boot.py ...